apps/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
sqlmodel>=0.0.22,<1.0
sqlalchemy>=2.0.30,<3.0
psycopg[binary]>=3.2,<4.0
PyJWT>=2.8,<3.0
python-dotenv>=1.0,<2.0
